feat: Adjustments to defaults, postgres docker example, minor fixes (#26)
* feat: Adjust interactive default prompts

They are similar when keeping everything, and when choosing to modify
and selecting all the default prompts (mainly changing gateway default)

* feat: Support custom app branch names with slash

Previously, the the app would fail to install due to file path issues,
now it works.

* Add docker postgres support via compose and update readme

Only for local testing if you don't want to install postgres binaries.

* Remove global npm install

- Installing globally doesn't work with how the daemon start is currently implemented. Rather than fix that, prefer a local install.
We don't need a global install. The test app brings its own node_modules folders and scripts to run ceramic/composedb, and this makes it easier to compare multiple versions.